requirements.txt
polygon-api-client==1.14.4
pandas== 2.2.3
python-dotenv==1.0.1
